+++ audio/wav/data/8bit.go
@@ -78,7 +78,7 @@ func (d *Chunk8bit) Generate(waveType osc.Type, freq, sampleRate int, dur time.D
 	if fn == nil {
 		return
 	}
-	fn(buffer, float64(freq), float64(d.Depth), float64(sampleRate))
+	fn(buffer, freq, int(d.Depth), sampleRate)
 
 	if d.Data == nil {
 		d.Data = buffer
